feat: :sparkles: Exit status stored in env
- env_set_exitcode()로 값 설정
- env_get()을 이용, 다른 환경 변수와 같은 방식으로 값 획득
