+++ makefile
@@ -25,7 +25,7 @@ parserV  := parser new1 new2 del util1 util2 expander
 promptV  := prompt interrupt util
 apiV     := shell signal path file
 execV    := context exec pipe redirect argv util
-builtinV := env util
+builtinV := env1 env2 util
 treeV    := repr1 repr2
 
 # ===== Macros =====
